chore: activate strict null checks

+++ libs/ngdux/util/store/form/src/lib/+state/abstract-form-effects.ts
@@ -5,14 +5,14 @@ import { of } from 'rxjs';
 import { catchError, exhaustMap, map, switchMap, tap } from 'rxjs/operators';
 import { FormActions } from '../models/form.model';
 
-export abstract class AbstractFormEffects<T, E> {
+export abstract class AbstractFormEffects<DTO, ERROR, CREATE_DTO = DTO> {
   load$ = createEffect(() =>
     this.actions$.pipe(
       ofType(this.formActions.load),
       switchMap(({ id }) =>
         this.service.loadResource(id).pipe(
           map(resource => this.formActions.loadSuccess({ resource })),
-          catchError((errors: E) => of(this.formActions.loadFailure({ errors })))
+          catchError((errors: ERROR) => of(this.formActions.loadFailure({ errors })))
         )
       )
     )
@@ -21,36 +21,48 @@ export abstract class AbstractFormEffects<T, E> {
   create$ = createEffect(() =>
     this.actions$.pipe(
       ofType(this.formActions.create),
-      switchMap(action =>
-        this.service.createResource(action.resource).pipe(
+      switchMap(action => {
+        if (!this.service.createResource) {
+          throw new Error('createResource not implement in the ListService');
+        }
+
+        return this.service.createResource(action.resource).pipe(
           map(resource => this.formActions.createSuccess({ resource })),
-          catchError((errors: E) => of(this.formActions.createFailure({ errors })))
-        )
-      )
+          catchError((errors: ERROR) => of(this.formActions.createFailure({ errors })))
+        );
+      })
     )
   );
 
   update$ = createEffect(() =>
     this.actions$.pipe(
       ofType(this.formActions.save),
-      switchMap(({ resource }) =>
-        this.service.saveResource(resource).pipe(
+      switchMap(({ resource }) => {
+        if (!this.service.saveResource) {
+          throw new Error('saveResource not implement in the ListService');
+        }
+
+        return this.service.saveResource(resource).pipe(
           map(response => this.formActions.saveSuccess({ resource: response })),
-          catchError((errors: E) => of(this.formActions.saveFailure({ errors })))
-        )
-      )
+          catchError((errors: ERROR) => of(this.formActions.saveFailure({ errors })))
+        );
+      })
     )
   );
 
   delete$ = createEffect(() =>
     this.actions$.pipe(
       ofType(this.formActions.delete),
-      exhaustMap(action =>
-        this.service.deleteResource(action.id).pipe(
+      exhaustMap(action => {
+        if (!this.service.deleteResource) {
+          throw new Error('deleteResource not implement in the ListService');
+        }
+
+        return this.service.deleteResource(action.id).pipe(
           map(() => this.formActions.deleteSuccess({ id: action.id })),
-          catchError((errors: E) => of(this.formActions.deleteFailure({ errors })))
-        )
-      )
+          catchError((errors: ERROR) => of(this.formActions.deleteFailure({ errors })))
+        );
+      })
     )
   );
 
@@ -73,8 +85,8 @@ export abstract class AbstractFormEffects<T, E> {
   protected constructor(
     protected readonly actions$: Actions,
     protected readonly store: Store,
-    private readonly service: FormService<T>,
-    private readonly formActions: FormActions<T, E>,
-    private readonly notificationService: FormNotificationService<E>
+    private readonly service: FormService<DTO, CREATE_DTO>,
+    private readonly formActions: FormActions<DTO, ERROR, CREATE_DTO>,
+    private readonly notificationService: FormNotificationService<ERROR>
   ) {}
 }
